Refactor: move project to src/ and reorganise into dashboard-centric structure
