+++ src/app/(dashboard)/[leagueID]/chips/ChipsTable.tsx
