+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
-import { ThemeProvider } from "@/components/theme-provider"; // ✅ import ThemeProvider
+import { ThemeProvider } from "@/components/common/ThemeProvider"; // ✅ import ThemeProvider
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
