+++ src/components/ui/responsive-info-card.tsx
