+++ tailwind.config.ts
@@ -1,10 +1,12 @@
 import type { Config } from "tailwindcss";
 
 const config: Config = {
-  darkMode: "class", // ✅ correct for Tailwind v4
+  darkMode: "class",
   content: [
-    "./app/**/*.{ts,tsx}",
-    "./components/**/*.{ts,tsx}",
+    "./src/app/**/*.{js,ts,jsx,tsx,mdx}",
+    "./src/components/**/*.{js,ts,jsx,tsx,mdx}",
+    "./src/features/**/*.{js,ts,jsx,tsx,mdx}", // NEW
+    "./src/lib/**/*.{js,ts,jsx,tsx,mdx}",      // optional, in case you use classes in lib
   ],
   theme: {
     extend: {},
